Code refactor and clean up
