+++ react/src/features/picturePicker/ImageDisplay.tsx
@@ -1,48 +1,33 @@
 import * as React from "react";
 import { useState, useEffect, useRef } from "react";
-import Modal from "./modal/modal";
-import { Celebrity } from "./modal/facedisplay";
-import { celebrities } from "./modal/facedisplay";
-// import { convertUrlToBase64 } from "./converturl";
-import { resolve } from "node:path/win32";
-import { rejects } from "node:assert";
-import Dropzone from "./modal/dropzone";
+import Modal from "./modal/Modal";
+import { ICelebrity } from "./FaceDisplay";
+import { celebrities } from "./FaceDisplay";
 
-export default function FirstComponent() {
-  // const [changePhoto, setChangePhoto] = useState("");
-  const [celeb, setCeleb] = useState<Celebrity>(celebrities[0]);
-  const [urlTo64, setUrlTo64] = useState("");
+export const ImageDisplay = () => {
+  const [celeb, setCeleb] = useState<ICelebrity>(celebrities[0]);
   const [image, setImage] = useState(celebrities[0]);
   const isMounted = useRef(false);
   const [file, setFile] = useState<File | null>(null);
   const [preview, setPreview] = useState<string | null>(null);
 
-  const handleClick = (celebrity: Celebrity) => {
+  const handleClick = (celebrity: ICelebrity) => {
     console.log(celebrity.id, celebrity.img);
     setCeleb(celebrity);
     console.log("Image clicked");
   };
 
-  // const urlImageCallback = (
-  //   event :  FormEvent<HTMLInputElement> | undefined
-  // ) : void => {
-  //   convertUrlToBase64(event, setUrlTo64);
-  // };
-
   const ANALYZE_POST_URL = "http://127.0.0.1:8000/verify";
 
-  //function signiture and then function call
   const urlToBase64 = (url: string) =>
     fetch(url)
       .then((response) => response.blob())
-      //convert image in memory into blob
       .then(
         (blob) =>
           new Promise((resolve, reject) => {
             const reader = new FileReader();
             reader.onloadend = () => resolve(reader.result);
             reader.onerror = reject;
-            //dataurl converts to base64
             reader.readAsDataURL(blob);
           })
       );
@@ -50,7 +35,6 @@ export default function FirstComponent() {
   useEffect(() => {
     if (isMounted.current) {
       urlToBase64(celeb.img).then((base64) => {
-        // console.log("Result:", base64);
         fetch(ANALYZE_POST_URL, {
           method: "POST",
           headers: { "Content-Type": "application/json" },
@@ -71,10 +55,9 @@ export default function FirstComponent() {
     } else {
       isMounted.current = true;
     }
-  }, [celeb]); //second argument is depedency array , when ever variable is change then useeffect is triggered
+  }, [celeb]);
 
   useEffect(() => {
-    //run when [file] changes
     const reader = new FileReader();
 
     reader.addEventListener(
@@ -89,7 +72,6 @@ export default function FirstComponent() {
       reader.readAsDataURL(file);
     }
   }, [file]);
-  //set the reult of the reader as my preview value
 
   return (
     <div>
@@ -101,4 +83,4 @@ export default function FirstComponent() {
       <Modal onImageSelect={handleClick} />
     </div>
   );
-}
+};
