+++ react/src/features/picturePicker/dropzoneImages/doc.png
