+++ react/src/features/picturePicker/modal/modal.css
